Create simple-calculator
Well, this exercise did get very complicated. The exercise asked to make a calculator by which at first a value was asked, then the operator had to enter it and finally the other value was asked. My mistake was not knowing how to use the logical operators, at the end is the code with the correct solution. I feel a little bad that I couldn't solve it correctly.